+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Node attribute expressions are now correctly reported as not allowed in
   resource defaults rules ([rhbz#1896458])
 - Upgreded to jquery 3.6.0 ([rhbz#1882291, rhbz#1886342])
+- Man page and help: note that 'pcs resource unclone' accepts clone resources
+  as well ([rhbz#1930886])
 
 ### Changed
 - python3-openssl was replaced with python3-cryptography ([rhbz#1927404])
@@ -25,6 +27,7 @@
 [rhbz#1886342]: https://bugzilla.redhat.com/show_bug.cgi?id=1886342
 [rhbz#1896458]: https://bugzilla.redhat.com/show_bug.cgi?id=1896458
 [rhbz#1927404]: https://bugzilla.redhat.com/show_bug.cgi?id=1927404
+[rhbz#1930886]: https://bugzilla.redhat.com/show_bug.cgi?id=1930886
 
 ## [0.10.8] - 2021-02-01